ai_workflow_builder/requirements.txt
pyside6>=6.0.0
litellm>=1.0.0
faiss-cpu>=1.7.0
docker>=6.0.0
fastapi>=0.95.0
pydantic>=2.0.0
watchdog>=3.0.0
nodegraphqt>=0.5.0
uvicorn>=0.22.0
numpy>=1.20.0
aiohttp>=3.8.0